API/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==5.0.0
beanie==2.0.1
boto3==1.42.17
botocore==1.42.17
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.127.0
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
jmespath==1.0.1
lazy-model==0.4.0
motor==3.7.1
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.15.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
PyYAML==6.0.3
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
watchfiles==1.1.1
websockets==15.0.1
